requirements.txt
tensorflow==2.12.0
keras==2.12.0
numpy==1.23.5
Pillow==9.5.0
matplotlib==3.7.1
scikit-learn==1.2.2
visualkeras==0.0.2
